feat(quizReducer): Export InitialState and ActionType types
Export the InitialState and ActionType types from the quizReducer file. This
allows other files to import and use these types. The InitialState type
defines the structure of the quiz state, including the currentQuestionIndex,
selectedOption, isSubmit, and score. The ActionType type defines the possible
actions that can be dispatched to the quizReducer, including SELECT_OPTION,
SUBMIT_ANSWER, and NEXT_QUESTION.

Issue: main