feat(hooks): Add showError property to quizReducer (main)
Add showError property to the quizReducer state interface. This property
will be used to determine whether or not to display an error message.
If the selected option is empty when submitting an answer, showError
will be set to true. Otherwise, it will be set to false.

Changes:
- Added showError property to QuizState interface
- Updated SUBMIT_ANSWER case in quizReducer to handle showError

Issue: N/A